Created some basic underlying architecture
An ingredient is a basic food element that can be sold out, such as
pulled pork, cheese, lettuce, bread, etc.

A MenuItem contains one or more ingredients. If any ingredient used in
the MenuItem is sold out, the whole thing will be considered sold out. A
MenuItem also keeps track of what it costs, which will be useful when
planning orders later.

These two classes work together so that at the restaurant when we sell
out of an ingredient we can press one button which will automatically
push changes to all affected menu items. Customers, however, will see
the menu in a way that is easy to read for them.