Deleted Android Stuff
It was causing problems with my Eclipse stuff, and the code somehow
ended up in two places at once. It's all in one central location now. We
can put the functionality in an Android project later but for now we
don't really need it. We weren't doing anything with it yet anyway.