Menu Fixes
The Menu used to be bugged when reading from json, because each menuitem
would get a list of unique ingredients, which wouldn't be changed when
they were supposed to. I have removed most of the OO design from the
package in favor of making it work easily with the gson library. Menu is
now responsible for figuring out if a menuitem is sold out or not.

Also added some more convenience methods to WebOut