Added a way to get/set the sold out password
We don't want users to have to enter a password each time they change
the sold out status, so I built in a way to read an FTP password without
just storing it in plain text somewhere. I might move this functionality
to its own class in the future, also I want to try to obfuscate this
further to make it harder for a hypothetical attacker to abuse.

I also added a new boolean to EncryptedFile to help tell if it has been
set up correctly.